+++ haegerp/src/main/java/org/haegerp/entity/repository/BusinessPartnerRepository.java
@@ -3,6 +3,6 @@ package org.haegerp.entity.repository;
 import org.haegerp.entity.BusinessPartner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusinessPartnerRepository extends JpaRepository<BusinessPartner, Integer> {
+public interface BusinessPartnerRepository extends JpaRepository<BusinessPartner, Long> {
 
 }
