+++ haegerp/src/main/java/org/haegerp/entity/repository/SalaryCategoryRepository.java
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional(readOnly=true, propagation=Propagation.MANDATORY)
-public interface SalaryCategoryRepository extends JpaRepository<SalaryCategory, Integer> {
+public interface SalaryCategoryRepository extends JpaRepository<SalaryCategory, Long> {
 
 	@Modifying
 	@Transactional
